Moves pulling of blog posts from instance initializer to application route and model hook
Fixes issue where blog posts weren't in FastBoot render, but were in the shoebox, and
therefore loaded on the browser

deferRendering on the fastboot service wasn't working to defer until the
blog posts were retrieved and in the shoebox before the application route
model hook was called. This makes sense as the route is not responsible for
rendering, but fastboot would still continue to render the application controller
template without the shoebox loaded. The shoebox would be loaded before the
response was sent to the browser however.